resolves #69
Added explicit fields for each of the CoRB2 options, the ability to set
project properties (with naming convention to prevent conflicts),
and/or System properties. CoRB options are now set using System
properties, allowing for any/all of the options to be specified when
executing the CorbTask.

Project properties use the naming convention of a “corb” prefix and
CamelCased name of the option (i.e. -PcorbThreadCount would set
THREAD-COUNT).

System properties will look for CoRB options with their original names
(i.e. -DTHREAD-COUNT).

Previous CorbTask member variables and behavior were preserved, such as
the convention of constructing the URIS-MODULE and PROCESS-MODULE if
modulePrefix value is specified.

Methods to collect and overlay the CoRB options are extracted into
methods that can be overridden if the user would like to change the
precedence or behavior.

Signed-off-by: Mads Hansen <mads.hansen@marklogic.com>