remove hard-coded member variables, generate and add through CorbTask.metaClass
Rather than have to maintain a set of member variables that mirror the
CoRB2 options, dynamically create and set them through the metaClass.
When the next version of CoRB2 is released, the current hard-coded list
of options can be replaced with a call to obtain the list of options
via reflection on the Options class. In this way, the CorbTask will
always reflect the current set of options.

Signed-off-by: Mads Hansen <mads.hansen@marklogic.com>