mlCopyRoxyFiles should do the following:
Accept a "-Ppath=/path/to/roxy/project" command-line property to specify where the Roxy project is located
Copy the contents of $path/src to ./src/main/ml-modules/root
Copy the contents of $path/rest-api/config to ./src/main/ml-modules/options
Copy the contents of $path/rest-api/ext to ./src/main/ml-modules/services
Copy the contents of $path/rest-api/transforms to ./src/main/ml-modules/transforms
Make a backup copy of gradle.properties if it exists - e.g. gradle.properties.backup. See NewProjectTask for an example of doing this.
Log each of the items above via println in the Gradle task so that the user has a nice record of everything that the task did.