#428 Now setting the Gradle projectDir on the AppConfig object
This allows ml-gradle to work with Java 11 and the Gradle daemon, as paths in AppConfig are now constructed relative to the projectDir.

While testing CreateResource/CreateTransform, also made a little fix so that the last module path is used when writing out a file. This avoids a problem where the files could be written to e.g. mlRestApi paths.