#334 Fixing some bugs with combined CMA requests
Turns out amps can't be included with databases and forests because REST API instances are created after them, but amps may depend on a modules database created by a REST API instance.

Also removed custom forests from a combined request, as there was little improvement to that, and it could lead to a bug in DHF with its custom command for deploying database fields.