Test Results for Nested Test Suites
Fixing a bug when writing test result XML for nested test
suites.  Originally the '/' or '\' in the suite path would
be treated as a separate directory by Groovy, which was
throwing file not found errors as those directories did not
exist on the local file system.  Now the task will replace
directory separators with '_' when writing the XUnit XML.