Escape Filename Tests
Adding a few automated test cases to show the issue that is being
addressed.  Gradle was throwing an error when it attempted to write to
filenames that included \ and / in the filename.  Those characters were
being treated as new directories when gradle was writing the results to
the file system.  The directory separators will be replaced with .
which matches the test result filenames used by Java and Groovy tests.