#422 DB file is passed to command for sub database
There's no change to the test, as this wasn't a bug in ml-app-deployer, but rather is helping avoid a bug in DHF until DHF can be fixed. It was really an oversight not to pass the subDatabaseFile in the first place, even though it's set on the command a few lines later.