Throwing error when requested command does not exist
This would only happen on a programming error - in which case we almost certainly want an error because the caller is expecting a command to be executed and that's not being done. Ran into this when I was testing out a PR but I had not yet published ml-app-deployer, and a task was silently failing because the command wasn't found. An error is much better in that scenario.