Improving handling of deleting a REST API server
Unfortunately the scenario in question - when a user has changed their modules-database to the filesystem - cannot be fully supported yet due to a server bug. But this at least exposes a better error to the user, which is currently being obscured by the error due to the invalid MIME type from the call to DELETE /v1/rest-apis.