Updated ssl-2way to work with ml-gradle 4.x
The previous version used ml-gradle 3.x and Java Client 4. With ml-gradle 4.x using Java Client 5 and higher, an X509TrustManager must be explicitly provided (has to do with the Java Client changing from using Apache HTTP to OkHttp between major versions 4 and 5). Setting `restTrustManager` accomplishes that. 

Also had to bump up the SSL version to the standard TLSv1.2.

This project will receive more updates soon so that it is easier to test with.