Added test app
This isn't perfect, and the CONTRIB needs a lot more work - but it's a start, and it lets us write more complete tests that don't depend on the admin user for everything. That'll be a future PR - trying to convert as many tests as possible to use a non-admin user.