DEVEXP-565 Using CMA to create forest replicas
There aren't any tests to add as the test cluster only has a single node. Instead doing manual testing against a 3-node cluster in Docker. 

The change is straightforward: if CMA is enabled, then submit a CMA config containing each of these forests with each of its replicas. If that fails, or if CMA is not enabled, then submit each forest one-by-one to /manage/v2.