Added properties for 2-way SSL
Added 8 properties for 4 connections - rest, appServices, manage, and admin - and then 8 more "global" properties that affect all 4 connections. 

The tests ensure that the properties are applied the config objects correctly, and the Java Client project has tests to verify that the properties work. Manual testing will be done with ml-gradle to ensure that everything works together.