Updated Gradle config
With the update to the Gradle publishing plugin, something changed such that declaring `artifact sourcesJar` and `artifact javadocJar` caused Gradle to complain about multiple artifacts with the same extension. They appear unnecessary now, so I removed them, so that local publishing works again. Will revisit when it's time to publish to Maven Central.