MLE-20741 Forest replica creation now considers host zone
This isn't fully baked yet, but it gets the plumbing in place for accounting for a zone value on each host. ForestReplicaPlanner was created almost entirely by Copilot, and it's handling the work of determining which replicas to build and on what hosts.

Next step will be to enhance DeployForestsCommand to actually build the map of hosts to zones, which will then enable the feature.

A final step, and in a separate ticket, will be to figure out what to do with ConfigureForestReplicasCommand. That hopefully is no longer needed, given that DeployForestsCommand is able to create replicas as well.