MLE-20741 Now accounting for zones when creating forest replicas
This hooks up the host-to-zone mapping so that it's fed into the forest replica planner. Currently only being tested via manual testing, which is consistent with how "real" replicas have been created in the past.

Enhanced the docker-compose file to create a 6-host cluster for manual testing purposes.

Story isn't quite done though. Need to add some tests for creating forests on a single host, as that seems a bit off.