MLE-22812 Refactor: Only configuring replicas in one place
Removed replica creation from DeployForestsCommand. ForestPlanner now handles both forests and replicas for the sake of previewing forest creation.

There's a lot more that can be done here - namely, in moving logic out of the command classes into a "forest calculator" class that handles all the logic for building and optionally saving a set of forests. But this at least gets rid of the two paths for replica creation.