MLE-23174 Using retry interceptor to fix intermittent failures
Another fix is to handle the exception that can handle due to MLE-547.

I am going to follow this up with another PR that may move the retry interceptor to src/test/java, as I'm not sure yet that we want to expose this to users.

Main goal right now is getting stable builds with no test failures.