MLE-23174 Hiding retry capability for now
This really needs to be supported in the Java Client in order for it to be done correctly - specifically in the DatabaseClientBuilder class. Otherwise we can only support it in here for the Manage/Admin connections but not the REST / App-Services connections.

Had to move TestConfig so that it could invoke the protected method for specifying a ClientConfigurator, which thus keeps it out of the public API.